+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Fixed
-- Tax query operator AND
+- Tax query fixes to support relation and operator simultaneously
 
 
 ## [1.16.3] - 2023-02-15