+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.0.7] - 2023-05-12
+
+### Fixed
+- Fixes tax_query `AND` operator.
+
 ## [2.0.6] - 2023-04-27
 
 ### Fixed